Quick-and-dirty fudge the 403
By going back to manually viewing the page in a fucking GUI browser.
Oh well.
I'm not really supposed to be consuming steamdb.info's downstream page like this ANYWAY.
I'm supposed to get a steam API key and talk to steam directly.
I am/was limiting this query to once-per-week maximum, which I figured is/was acceptable.